Moved state machine test task and clean up.
