+++ VCU/.settings/org.eclipse.core.resources.prefs
@@ -27,6 +27,8 @@ encoding//BuildConfig_Launchpad-1224/Phantom/tasks/subdir_rules.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/Phantom/tasks/subdir_vars.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/Phantom/test_eeprom/subdir_rules.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/Phantom/test_eeprom/subdir_vars.mk=UTF-8
+encoding//BuildConfig_Launchpad-1224/Phantom/test_statemachine/subdir_rules.mk=UTF-8
+encoding//BuildConfig_Launchpad-1224/Phantom/test_statemachine/subdir_vars.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/halcogen_launchpad/source/subdir_rules.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/halcogen_launchpad/source/subdir_vars.mk=UTF-8
 encoding//BuildConfig_Launchpad-1224/makefile=UTF-8
