+++ VCU/Phantom/tasks/source/task_throttle.c
@@ -126,18 +126,18 @@ void vThrottleTask(void *pvParameters){
             if (BSE_sensor_sum < BSE_MIN_VALUE)
             {
                 // if it's less than 0.5V, then assume shorted to GND as this is not normal range
-                VCUDataPtr->DigitalVal.BSE_FAULT = 1;
+                VCUDataPtr->DigitalVal.BSE_SEVERE_RANGE_FAULT = 1;
 
             }
             else if (BSE_sensor_sum > BSE_MAX_VALUE) // change from magic number to a #define BSE_MAX_VALUE
             {
                 // if it's greater than 4.5V, then assume shorted to 5V as this is not normal range
-                VCUDataPtr->DigitalVal.BSE_FAULT = 1;
+                VCUDataPtr->DigitalVal.BSE_SEVERE_RANGE_FAULT = 1;
             }
             else
             {
                 // should be in normal range
-                VCUDataPtr->DigitalVal.BSE_FAULT = 0;
+                VCUDataPtr->DigitalVal.BSE_SEVERE_RANGE_FAULT = 0;
             }
 
             // moving average signal conditioning.. worth it to graph this out and find a good filter time constant
